Override Fallback Region A user-selected fallback to use instead of the default PAL
This can be used for unknown region or region free titles to give them
the ability to force a different region than the current fallback region
of PAL. This can be useful if a user is trying to play a region free
tilte that is originally NTSC and expects to be run at NTSC speeds. This
may be done when a user attempts to dump a WAD of their own without
understanding the settings they have chosen, or could be an intentional
decision by a developer of a ROM hack that can be injected into a
Virtual Console WAD.

This is primarily a proof-of-concept in its current stage.